Functional for depths up to 16

+++ packboost/cuda/repack.cu
@@ -8,7 +8,7 @@ using at::Tensor;
 
 // FST:   [nsets, nfeatsets, max_depth], uint8
 // L_old: [nfolds, Dm, N],                uint8 branch bits
-// LF:    [nfeatsets, Dm, N],             uint8 depth-local prefixes
+// LF:    [nfeatsets, Dm, N],             uint16 depth-local prefixes
 // Logic: LF[fs, d, j] = prefix(d+1 bits) from fold FST[tree_set, fs, d+1]
 
 __device__ __forceinline__ int fst_idx(int k, int d, int lane, int Dm) {
@@ -19,7 +19,7 @@ __device__ __forceinline__ int fst_idx(int k, int d, int lane, int Dm) {
 __global__ void repack_trees_for_features_kernel(
     const uint8_t* __restrict__ FST,    // [nsets, nfeatsets, max_depth]
     const uint8_t* __restrict__ L_old,  // [nfolds, Dm, N]
-    uint8_t* __restrict__ LF,           // [nfeatsets, Dm, N]
+    uint16_t* __restrict__ LF,          // [nfeatsets, Dm, N]
     int nsets, int nfeatsets, int max_depth, int nfolds, int Dm, int N,
     int tree_set, int stride
 ){
@@ -57,10 +57,10 @@ __global__ void repack_trees_for_features_kernel(
                 const int fs = 8 * fi + k;
                 if (fs < nfeatsets) {
                     const uint32_t which_fold = fst[fst_idx(k, d, wi, Dm)];
-                    uint8_t prefix = 0;
+                    uint16_t prefix = 0;
                     for (int i = 0; i <= d; ++i) {
                         const uint8_t b = L_old[((which_fold * Dm) + i) * N + j];
-                        prefix = static_cast<uint8_t>((prefix << 1) | (b & 1u));
+                        prefix = static_cast<uint16_t>((prefix << 1) | (b & 1u));
                     }
                     LF[((fs * Dm) + d) * N + j] = prefix;
                 }
@@ -73,7 +73,7 @@ __global__ void repack_trees_for_features_kernel(
 void repack_trees_for_features_cuda(
     const torch::Tensor& FST,      // uint8 [nsets, nfeatsets, max_depth]
     const torch::Tensor& L_old,    // uint8 [nfolds, Dm, N]
-    torch::Tensor& LF,             // uint8 [nfeatsets, Dm, N] (output)
+    torch::Tensor& LF,             // uint16 [nfeatsets, Dm, N] (output)
     int64_t tree_set               // which set (round)
 ) {
     const int nsets      = static_cast<int>(FST.size(0));
@@ -101,7 +101,7 @@ void repack_trees_for_features_cuda(
         <<<grid, block, smem, stream>>>(
             FST.data_ptr<uint8_t>(),
             L_old.data_ptr<uint8_t>(),
-            LF.data_ptr<uint8_t>(),
+            LF.data_ptr<uint16_t>(),
             nsets, nfeatsets, max_depth, nfolds, Dm, N,
             static_cast<int>(tree_set),
             inner_stride
